splitting up decrypter into multiple files
