+++ src/decrypter/index.js
@@ -1,0 +1,9 @@
+import decrypt from './decrypt';
+import AsyncStream from './async-stream';
+import Decrypter from './decrypter';
+
+export default {
+  decrypt,
+  Decrypter,
+  AsyncStream
+};
